requirements.txt
# core dependencies
PyQt5>=5.15.0
python-dotenv>=0.19.0
openai>=1.0.0
Pillow>=8.0.0

# tts dependencies
pyttsx3>=2.90
pygame>=2.1.0  # for audio playback with OpenAI TTS

# database
sqlite4>=0.0.1  # check version with your installation

# testing and development
pytest>=7.0.0
flake8>=4.0.0

# arduino
simpleaudio
pyserial